Added phone item and phone call to basic game
I extended the basic IslandGame to include an inventory and a phone item
that the player can pick up and drop. If the player has the phone in their
inventory, they will sometimes recieve notifications from the phone in
terms of dialogue modals.

I am workign towards the goal of slowly increase the probability of reciving a notification every time you use your phone.

Estimate for time remaining to finish assignment: 2 hours(