Added dialogue, fixed game over bug
I added more types of notifications and dialogue interactions for them. I fixed a bug where the game over message was displaying on the original center of the screen rather than the center of where the view port was currently focused.

I'm now done with the basic part of my game.